Modify headless scanner/scanner opts for arg-class option
Requires the Soot OptionBase method
setArgClass. Currently there is no way for
an application calling Soot via the Scene
to specifiy only one class to process,
this way we can control this. Previously had to
rely on file system management to assure that
a process-dir had only one file in it if we
desired this use case, and we do not always have control of
this.

Signed-off-by: knewbury <knewbury@ualberta.ca>